Generate binary which handles 17/22 TPC-H queries
Works on small dataset. Requires a few minor manual tweaks post
generation, to get things like filenames and onion levels right.  While
this is sub-optimal, it is feasible for now. There is no need to tweak
any of the plan node operators.

This forms a stable baseline to build off of