Dictionary compression for sending batches of values back
This commit attempts to gain back some of the perf losses over a
throttled network. The idea is that for deterministic columns which have
a low number of distinct values (ie the l_quantity column in the
lineitem table), it is wasteful to send full ciphertexts back.

Instead, the client asks the server stats table for the most
frequent ciphertext values for that column, and keeps all values
above a certain threshold (0.02 in the current implementation).
The client then informs the server of this dictionary, and
when running queries, asks the server to use this dictionary
when sending group-ed data back.

So for instance, suppose we do originally:

  select group_serialize(l_quantity_DET) from ...
  group by l_orderkey_DET

Before, each l_orderkey_DET group will have N*8 bytes for
each of the N ciphertexts in the group.

Now, we tell the server:

  select group_serialize(l_quantity_DET, [dictionary-id]) from ...
  group by l_orderkey_DET

And the resulting value for each group is the following format.
Let N_DictHits = number of dictionary hits.

  N_DictHits ||
  [N_DictHits variable-length uint32 indices into dict table] ||
  [Delta encoded uint64 array for the remaining ciphertexts]

By implementing this optimization, we are able to shave off
30 seconds of runtime for Q18 over a bandwidth throttled
network at 10mbits/sec.