Add the ability to toggle multi-pos multiplications
Also add the ability to really disable debug stuff in the aggs

For some reason our implementation of multi-pos muplications
is not giving us the expected performance gains. In fact
initial testing suggests it might even be slower.

Not sure why this is the case, but we can always just disable
this optimization going forward.